Changed index.html to have log-in forms.
I'm not doing anything to it at this point. Also deleted some useless CSS. Added comments to lines 14 and 17 on SignUpController.java.

Really need to figure out how to create a story via the application, or rather pass the parameters correctly to the postmapping of create() in StoryController.

Also need to make sign-up work.