Summary: Finally solved the problem with creating a story in the application.
Problem Was Caused By:
I noticed that the problem wasn't with the backend, because I could go into ptpython and use the route from there. After noticing that, I deduced that the problem was with the javascript itself. I was unaware that you have to declare that you have headers in your fetch when you post.

Problem Was Solved By:
I added headers to lines 211-213 to index.js. The reason for headers is to tell your program what to read the fetch contents as. Otherwise you get a 415 HTTP response.

Need to do:
I've gotten create done, now I need to make signup work.