Summary: Fixed small bugs with login.
Problem:
DB data would still be shown in the browser, because it was just hidden.

Solved By:
Changing the js so it wouldn't load the DB until the user has been logged in.  I made it to where the stories and users did not load when the window was pulled up.

Need To Do:
Logout, plus a short list of bugs I found.